Update Hakase to use Discord.js v12
Major revamp as Discord.js v12 introduces a lot of breaking changes,
including API changes and use of cache instead of direct collections. In
particular, sendMessage is no longer available and instead the general
send() has to be used instead which is asynchronous, thus await is used to
keep most of the code changes minimal. Also:
- removed unused files
- add embed help command
- update readme documentation