Fix good/bad bot query
Allow good/bad bot query to be specified in one line as well. Also fixes an
bug that can occur where the user may call Hakase, someone else may say
something, then the original caller queries good/bad bot, but Hakase will not
respond. This happens because the code assumes that good/bad bot will always
be sent immediately after a Hakase message, but if lots of people are messaging
at the same time this may not be the case. To resolve this, we basically do the
same check again in the interpretHakaseQuery function as we do in init.js but
this time check if the previous messenger was not Hakase. This prevents a
double post by Hakase. This is also fine as the code path in
interpretHakaseQuery will only ever be reached when Hakase is called.