Fix missing engine in OpenAI query
For some reason, the content-filter-alpha engine string gets lost from
the global contentFilterParameters after the first content filter request.
To resolve this, we move the variables inside the function as they are
only ever used there anyway, and this way the correct parameters get
constructed every time. The ratelimit global variable is now attached
to the `global` object too.