Fix image sourcing to use whitelist instead of blacklist
The trouble with blacklist is that if the URL does not match the
blacklist then it can slip through. A whitelist is better to ensure
no unwanted sources gets allowed. There is a slight issue with
the Sagiri package in that the bitmask generation for anything
above 17 is wrong. This is rectified by using my own fork of the
package.