Add support for ImageKit to store users.json file
Implemented support for using ImageKit as an external storage host
for storing the users.json file. The CDN caches must be purged so
that we get the latest file upon every restart, however, we only
get 1000 caches purges a month. To be seen if this limit is reached.
ImageKit also uses promises, so a few functions have been converted
to use awaits. README also updated.